15/out

+++ Projeto/projeto/app/Http/Controllers/MusicaController.php
@@ -7,7 +7,10 @@ use App\Models\Musica;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
+
+
 class MusicaController extends Controller
 {
     /**
@@ -19,11 +22,26 @@ class MusicaController extends Controller
     {
 
         $musicas = Musica::orderBy('nome')->get();
-        $generos = Genero::orderBy('nome')->get();
-        
-        $artistas = DB::table('musicas')->distinct()->pluck('artista');        
 
-        return view('musicas.index',['musicas' => $musicas, 'generos' => $generos, 'artistas' =>$artistas]);
+        $generos = Genero::orderBy('nome')->get();
+
+        $artistas = DB::table('musicas')->distinct()->pluck('artista');     
+
+    
+        $lista = array();
+
+        foreach($musicas as $musica){
+            $generoAtual = $musica->genero->nome;
+
+            if ($generoAtual == Auth::user()->genero_fav){
+                array_push($lista, $musica);
+            }
+ 
+         }
+       
+
+        return view('musicas.index',['musicas' => $musicas, 'generos' => $generos, 'artistas' =>$artistas,
+                                    'lista' => $lista]);
     }
 
     /**
@@ -33,8 +51,14 @@ class MusicaController extends Controller
      */
     public function create()
     {
-        $generos = Genero::orderBy('nome')->get();
-        return view('musicas.create',['generos' => $generos]);
+        if(Auth::user()->tipo == 'A'){
+            $generos = Genero::orderBy('nome')->get();
+            return view('musicas.create',['generos' => $generos]);
+        }else{
+            session()->flash('mensagem', 'Infelizmente você não tem permissão :(');
+            return redirect()->route('welcome');
+        }
+
     }
 
     /**
@@ -45,50 +69,52 @@ class MusicaController extends Controller
      */
     public function store(Request $request)
     {       
-        //$idAtual = Musica::create($request->all());
+        if(Auth::user()->tipo == 'A'){
         
-        $idAtual = DB::table('musicas')->insertGetId(
-            ['nome' => $request->nome,
-            'artista' => $request->artista,
-            'album' => $request->album,
-            'genero_id' => $request->genero_id,
-            'ano' => $request->ano]
-        );
+            $idAtual = DB::table('musicas')->insertGetId(
+                ['nome' => $request->nome,
+                'artista' => $request->artista,
+                'album' => $request->album,
+                'genero_id' => $request->genero_id,
+                'ano' => $request->ano]
+            );
+
+        
+            //se informou arquivo
+            if($request->hasFile('caminho') && $request->file('caminho')->isValid()){
+            
+            
+                //dando nome ao arquivo
+                $name = Str::kebab($request->nome);
+                $name2 = Str::kebab($request->artista);            
+                $extension = $request->caminho->extension();
+                $nameFile = "{$name}-{$name2}.{$extension}";
+            
+
+                $upload = $request->caminho->storeAs('music', $nameFile);
+                
+                if(!$upload)
+                    return redirect()
+                                ->back()
+                                ->with('error', 'Falha ao realizar upload');    
+
+         }
 
        
-        //se informou arquivo
-        if($request->hasFile('caminho') && $request->file('caminho')->isValid()){
-           
-            //if($user->foto){
-              //  $name = $user->foto;
-            //}else{
-            //dando nome ao arquivo
-            $name = Str::kebab($request->nome);
-            $name2 = Str::kebab($request->artista);            
-            $extension = $request->caminho->extension();
-            $nameFile = "{$name}-{$name2}.{$extension}";
-            //dd($nameFile);
 
-            $upload = $request->caminho->storeAs('music', $nameFile);
-            
-            if(!$upload)
-                return redirect()
-                            ->back()
-                            ->with('error', 'Falha ao realizar upload');    
+            DB::table('musicas')
+                ->updateOrInsert(
+                    ['id' => $idAtual],
+                    ['caminho' => $nameFile]
+                
+            );
 
-       // }
-
+            session()->flash('mensagem', 'Musica cadastrada com sucesso!');
+            return redirect()->route('musicas.index');
+        }else{
+            session()->flash('mensagem', 'Infelizmente você não tem permissão :(');
+            return redirect()->route('welcome');
         }
-
-        DB::table('musicas')
-            ->updateOrInsert(
-                ['id' => $idAtual],
-                ['caminho' => $nameFile]
-            
-        );
-
-        session()->flash('mensagem', 'Musica cadastrada com sucesso!');
-        return redirect()->route('musicas.index');
     }
 
     /**
@@ -133,22 +159,26 @@ class MusicaController extends Controller
      */
     public function destroy(Musica $musica)
     {
-       
-    if ($musica->delete()) {
-        //Deleta o arquivo da musica)
-        Storage::delete("music/{$musica->caminho}"); // true ou false
- 
-      
-        return redirect()
-                    ->route('musicas.index')
-                    ->with('mensagem', 'Sucesso ao deletar!');
-    }
+        if(Auth::user()->tipo == 'A'){
+            if ($musica->delete()) {
+                //Deleta o arquivo da musica)
+                Storage::delete("music/{$musica->caminho}"); // true ou false
+        
+            
+                return redirect()
+                            ->route('musicas.index')
+                            ->with('mensagem', 'Sucesso ao deletar!');
+            }
 
-        return redirect()
-                ->back()
-                ->with('mensagem', 'Falha ao deletar!');
+                return redirect()
+                        ->back()
+                        ->with('mensagem', 'Falha ao deletar!');
 
 
 
+            }else{
+                session()->flash('mensagem', 'Infelizmente você não tem permissão :(');
+                return redirect()->route('welcome');
+            }
     }
 }
